requirements.txt
streamlit==1.28.1
openai==1.1.2
newsapi-python==0.2.7
bs4==0.0.1